change composer to function
it only performs a single calculation
